+++ src/Position.java
@@ -1,20 +1,24 @@
 /* Max Besley. May 2022. */
 
 /**
- * A mutable class that represents an (x, y) coordinate/position.
+ * A mutable class that represents an (x, y) coordinate.
  */
 public class Position {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
 
-    // Constructor
-    public Position(int x, int y) {
+    /*
+     * Creates a position with the specified x, y coordinates.
+     */
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    // Copy constructor
+    /*
+     * Copy constructor.
+     */
     public Position(Position aPosition) {
         if (aPosition == null) {
             System.err.println("Error: null reference passed to Position class's copy constructor");
@@ -24,19 +28,33 @@ public class Position {
         this.y = aPosition.y;
     }
 
-    // Getters
-    public int getX() {
+    /*
+     * Returns the x coordinate of a position.
+     */
+    public double getX() {
         return x;
     }
-    public int getY() {
+
+    /*
+     * Returns the y coordinate of a position.
+     */
+    public double getY() {
         return y;
     }
 
-    // Setters
-    public void setX(int x) {
+    /*
+     * Sets the x coordinate of a position
+     * to the value passed as a parameter.
+     */
+    public void setX(double x) {
         this.x = x;
     }
-    public void setY(int y) {
+
+    /*
+     * Sets the y coordinate of a position
+     * to the value passed as a parameter.
+     */
+    public void setY(double y) {
         this.y = y;
     }
 }
